+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "vstosc" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## vstosc 1.3.0
+
+- Added command [vstosc:testRFunction](README.md#testrfunction).
+- Improved command [vstosc:updateRDocstring](README.md#updaterdocstring). It now also works when the cursor is positioned anywhere inside the function or its current roxygen docstring.
+
 ## [Unreleased]
 
 - Initial release